data: threshold datatype actually float and change hsv scanner count_nonzero method label

+++ bot/general/components/world_to_screen/components/screen_scanners/template_scanner/scanner.py
@@ -33,7 +33,7 @@ class TemplateScanner:
         *compiled_templates: CompiledTemplate,
         iterable_templates: Iterable[CompiledTemplate] = list(),
         region: Region = None,
-        threshold: int = 1
+        threshold: float = 1
     ):
         assert 0 < threshold <= 1
